+++ admin-ui/src/component/ReservationUnits/DateTimeInput.tsx
@@ -1,12 +1,24 @@
 import React from "react";
 import { DateInput, TimeInput } from "hds-react";
 import { useTranslation } from "react-i18next";
-import {
-  dateTime,
-  valueForDateInput,
-  valueForTimeInput,
-} from "../../common/util";
+import { parse } from "date-fns";
 import { EditorColumns } from "./editorComponents";
+import { formatDate } from "../../common/util";
+
+/* Convert api datetime to date required by date input, defaults to current date */
+export const valueForDateInput = (from: string): string => {
+  return formatDate(from || new Date().toISOString(), "d.M.yyyy") as string;
+};
+
+/* Convert api datetime to time required by time input,m defaults to current time */
+export const valueForTimeInput = (from: string): string => {
+  return formatDate(from || new Date().toISOString(), "HH:mm") as string;
+};
+
+/* Construct date from dateinput + timeinput */
+export const dateTime = (date: string, time: string): string => {
+  return parse(`${date} ${time}`, "dd.MM.yyyy HH:mm", new Date()).toISOString();
+};
 
 const DateTimeInput = ({
   value,
